# HTTP/2 Flow Controller should use Channel.isWritable()