--- a/codec-http2/src/main/java/io/netty/handler/codec/http2/Http2FlowController.java
+++ b/codec-http2/src/main/java/io/netty/handler/codec/http2/Http2FlowController.java
@@ -20,6 +20,15 @@ import io.netty.channel.ChannelHandlerContext;
  * Base interface for all HTTP/2 flow controllers.
  */
 public interface Http2FlowController {
+    /**
+     * Set the {@link ChannelHandlerContext} for which to apply flow control on.
+     * <p>
+     * This <strong>must</strong> be called to properly initialize the {@link Http2FlowController}.
+     * Not calling this is considered a programming error.
+     * @param ctx The {@link ChannelHandlerContext} for which to apply flow control on.
+     * @throws Http2Exception if any protocol-related error occurred.
+     */
+    void channelHandlerContext(ChannelHandlerContext ctx) throws Http2Exception;
 
     /**
      * Sets the connection-wide initial flow control window and updates all stream windows (but not the connection
@@ -63,11 +72,10 @@ public interface Http2FlowController {
      * window size published by this endpoint. It is up to the implementation, however, as to when a
      * {@code WINDOW_UPDATE} is actually sent.
      *
-     * @param ctx The context for the calling handler
      * @param stream The subject stream. Use {@link Http2Connection#connectionStream()} for
      *            requesting the size of the connection window.
      * @param delta the change in size of the flow control window.
      * @throws Http2Exception thrown if a protocol-related error occurred.
      */
-    void incrementWindowSize(ChannelHandlerContext ctx, Http2Stream stream, int delta) throws Http2Exception;
+    void incrementWindowSize(Http2Stream stream, int delta) throws Http2Exception;
 }