@Scottmitch 2015-06-05
@nmittler @ejona86 - FYI. I removed the `ChannelHandlerContext` as arguments to the flow controller methods. Having `ChannelHandlerContext` as an argument to the flow controller methods doesn't make much sense because the flow controller is only designed to work for 1 connection.  The current interface makes it easy to potentially pass in different `ChannelHandlerContext` for each operation.  The flow controller implementations were also saving/validating it in some (not all) places, and using it later.  This single reference to a `ChannelHandlerContext` is then used for all flow controller activity.