@ejona86 2015-06-08
It seems like a high percentage of the time the flow controller would still write how much it used to, since `writePendingBytes()` will write everything that flow control permits, and the `isWritable()` check at the beginning of the method is likely true since it will be called from `writabilityChanged()`. It does let the queue drain before filling it again, but it doesn't seem to reduce the maximum amount queued. The batchy-ness may cause trouble with calculating the RTT if we ever do advanced flow control.

@Scottmitch, I had originally anticipated the fix to be invasive. So I had no expectation that you would fix it, since you likely wouldn't have had the time to invest. I figured Nathan or I would likely need to be the one to spend the time on it. I'm very happy that you've looked into it though :)