@Scottmitch 2015-06-08
@ejona86 - The granularity of checking here is currently relatively corse (1 check at each public API entry point) but can be increased now that we have a "reliable" `ChannelHandlerContext` set in the flow controller.  The PR as it stands now may be a good stepping stone to more granular checks, as these changes will prevent writes going through to the underlying layer at the granularity of the public API.  There may be a window of writes from where the public API is called to the writability changed that still are still passed to lower layers, but this seems like an improvement over current code where writability does not factor in.

To increase the granularity of checking for writability I think we may need to explore the following:

1. Currently the allocation and writing are done in two independent steps.  The writing step is not done in priority order and is currently done by iterating over all active streams to write the amount allocated in the previous iteration.  This works well if we assume all of the allocations will be written but if we are going to stop writing prematurely we should be writing in priority order.  I attempted to unify the send/allocate loop once before https://github.com/netty/netty/commit/c8a1d077b50952e0c8b4dde67ddc7fc8b98308b1 (and commit comment still indicates this), and also track writes in priority order, but ended up keeping them separate to simplify logic.
2. If we are going to go down to this granularity it may be worth adding additional benchmarking (you may have this for gRPC already) to understand the implications. 
3. The writability check does include a `volatile` variable read. If we do plan to check after each write operation we should attempt to minimize calls to this method...may be relatively minor in the scheme of things but worth considering.

Should we start with this PR and then come up with a followup PR?  I think this is my preference but WDYT?