@nmittler 2015-06-09
@Scottmitch I think we need to add something like `getWritableBytes` to `Channel` (and `ChannelOutboundBuffer`).  Maybe we could make a PR to do that first, then this PR would get simpler. WDYT?