@nmittler 2015-06-09
Actually it looks like `ChannelOutboundBuffer` already has `totalPendingWriteBytes` ... so between that and the writability flag we're probably good.