@nmittler 2015-06-09
@Scottmitch I do think we should probably proceed with the fine-grained writability testing that @ejona86  is suggesting. It may be something as simple as checking the number of bytes we can write at the beginning of `writePendingBytes` and then limiting the connection window based on that.