@Scottmitch 2015-06-09
@nmittler @ejona86 - 

> It may be something as simple as checking the number of bytes we can write at the beginning of writePendingBytes and then limiting the connection window based on that.

Yes I can do that in this PR.  Just FYI the check will be based upon how many bytes are queued in the `ChannelOutboundBuffer` vs the low/high water marks (bcz this is how writability is controlled).  So it is possible that a lower level write will return EAGAIN but isWritable will return true as long as the pending bytes is lower than the high water mark.  If you want to be aggressive about getting writability changes from the underlying transport (and minimize queuing in `ChannelOutboundBuffer`) you can set the low/high water marks to a low level (say 0).