@ejona86 2015-06-09
@Scottmitch, I think I'd prefer not actually observing `ChannelOutboundBuffer` currently queued amount. Instead, we just decide to write a "reasonable" amount and re-check isWritable after writing. The "reasonable" amount could be based on the low/high watermarks (e.g., 1/2 high water mark, or difference between high and low).

I'm suggesting this because 1) observing `ChannelOutboundBuffer` requires using unsafe and 2) I'm concerned about writing small amounts to `writePendingBytes()`. `writePendingBytes()` currently divides the number of bytes across all streams, and I'm worried about writing very small amounts for each stream thereby causing horrible goodput. Using something less dynamic, like the watermark configuration directly, would make the behavior more predictable and would reduce the risk.