@Scottmitch 2015-06-09
Just to summarize a discussion @nmittler and @ejona86 had offline. The first approach is going to limit the connection window based upon the bytes until writability changes for the Channel (or ChannelOutBoundBuffer), and taking into account some minimum factor so that we don't run the allocation algorithm with a value that is "too small".