@Scottmitch 2015-06-10
@nmittler @ejona86 - PTAL. I included the commit from PR https://github.com/netty/netty/pull/3879 just to make progress on this PR...so just ignore the Channel and AbstractChannel related changes for this PR. I basically capped the "minim usable channel bytes" at `1 << 14` but we can adjust this...or make it configurable if necessary.