@nmittler 2015-06-11
@Scottmitch I took another pass.  Just a couple of items that need addressing.