@nmittler 2015-06-11
@Scottmitch looks good! I would recommend that we hold off committing this until @louiscryan completes work on #3860, since these overlap and will affect each other's performance.  After he commits, let's rebase this PR ... I think it may get simpler. Sound reasonable?