@Scottmitch 2015-07-01
@nmittler @louiscryan - I rebased my changes. I will try to run some benchmarks tomorrow.